+++ ios/RNSpotifyRemoteAuth.h
@@ -14,6 +14,9 @@
 +(instancetype)sharedInstance;
 
 -(void)initialize:(NSDictionary*)options resolve:(RCTPromiseResolveBlock)resolve reject:(RCTPromiseRejectBlock)reject;
+-(void)endSession:(RCTPromiseResolveBlock)resolve reject:(RCTPromiseRejectBlock)reject;
+-(void)getSession:(RCTPromiseResolveBlock)resolve reject:(RCTPromiseRejectBlock)reject;
+
 
 -(SPTConfiguration*) configuration;
 -(NSString*) accessToken;