+++ src/SpotifyRemote.ts
@@ -85,6 +85,14 @@ export interface SpotifyRemoteApi extends TypedEventEmitter<SpotifyRemoteEvents>
     connect(accessToken: string): Promise<void>;
 
     /**
+     * Disconnects the App Remote 
+     *
+     * @returns {Promise<void>}
+     * @memberof SpotifyRemoteApi
+     */
+    disconnect(): Promise<void>;
+
+    /**
      * Play a track, album, playlist or artist via spotifyUri
      * Example: spotify:track:<id>, spotify:album:<id>, spotify:playlist:<id>, spotify:artist:<id>
      * @param {string} spotifyUri